Remove temporary variable used for swapping
