+++ 02_bubble/02_bubble.go
@@ -2,20 +2,21 @@ package main
 
 import (
 	"fmt"
+	"io"
+	"os"
 )
 
+var out io.Writer = os.Stdout
+
 func main() {
-	fmt.Println(bubble([]int{3, 4, 1, 5, 2}))
+	fmt.Fprintln(out, bubble([]int{3, 4, 1, 5, 2}))
 }
 
 func bubble(s []int) []int {
-	var temp = 0
 	for i := 1; i < len(s); i++ {
 		for j := 0; j < len(s)-1; j++ {
 			if s[j] > s[j+1] {
-				temp = s[j]
-				s[j] = s[j+1]
-				s[j+1] = temp
+				s[j], s[j+1] = s[j+1], s[j]
 			}
 		}
 	}
