+++ 02_bubble/02_bubble_test.go
@@ -1,6 +1,8 @@
 package main
 
 import (
+	"bytes"
+	"strconv"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -25,4 +27,18 @@ func TestBubbleSortOutputForNegativeValues(t *testing.T) {
 	actual := bubble([]int{20, 0, -10, 100, -5})
 	r.Equal(expected, actual)
 
+}
+func TestMainOutput(t *testing.T) {
+	// Given
+	r := require.New(t)
+	var buf bytes.Buffer
+	out = &buf
+
+	// When
+	main()
+
+	// Then
+	expected := strconv.Quote("[1 2 3 4 5]\n")
+	actual := strconv.Quote(buf.String())
+	r.Equalf(expected, actual, "Unexpected output in main()")
 }
